Renewed the daw message so it also sends sax and say

+++ x32/fixed_point.h
@@ -32,8 +32,8 @@
 #define MULT(A,B)  			((int)(A * B) >> FRACT_PART)
 #define MULT_S(A,B,SIZE) 	((int)(A * B) >> SIZE)
 
-#define DIV(A,B)  			((int)(((long) A << FRACT_PART)/B))
-#define DIV_S(A,B,SIZE)  	((int)(((long) A << SIZE)/B))
+#define DIV(A,B) 			 (int)(((A) << FRACT_PART)/(B))
+#define DIV_S(A,B,SIZE)  	((int)(( A << SIZE)/B))
 
 
 #endif
